+++ src/main/java/shadows/apotheosis/Apotheosis.java
@@ -75,6 +75,7 @@ public class Apotheosis {
 	public static boolean enableEnch = true;
 	public static boolean enablePotion = true;
 	public static boolean enableVillage = true;
+	public static boolean giveBook = true;
 
 	public static float localAtkStrength = 1;
 
@@ -87,6 +88,7 @@ public class Apotheosis {
 		//enableDeadly = config.getBoolean("Enable Deadly Module", "general", true, "If the deadly module is loaded.");
 		enablePotion = config.getBoolean("Enable Potion Module", "general", true, "If the potion module is loaded.");
 		enableVillage = config.getBoolean("Enable Village Module", "general", true, "If the village module is loaded.");
+		giveBook = config.getBoolean("Give Book on First Join", "general", true, "If the Chronicle of Shadows is given to new players.");
 		config.setTitle("Apotheosis Module Control");
 		config.setComment("This file allows individual modules of Apotheosis to be enabled or disabled.\nChanges will have no effect until the next game restart.");
 		if (config.hasChanged()) config.save();